requirements.txt
# Flask for the web server
flask

# PyTorch and computer vision
torch
torchvision

# Image processing and numerical operations
numpy
Pillow

# For SAM and image processing
segment-anything
opencv-python-headless

# For YOLOv8 object detection
ultralytics

# For local VLMs (Hugging Face)
transformers
accelerate
bitsandbytes

# Required by Qwen
einops
tiktoken
qwen-vl-utils

# 8 bit quntization
bitsandbytes